+++ core/tests/test_ansible_manager.py
@@ -1,0 +1,1 @@
+__author__ = 'mike'
